World Happiness Report
The World Happiness Report is a landmark survey of the state of global happiness. The first report was published in 2012, the second in 2013, the third in 2015, and the fourth in the 2016 Update. The World Happiness 2017, which ranks 155 countries by their happiness levels, was released at the United Nations at an event celebrating International Day of Happiness on March 20th.

The happiness scores and rankings use data from the Gallup World Poll. The scores are based on answers to the main life evaluation question asked in the poll. 

You can answer the following questions when performing an analysis on this dataset:

Which country is the happiest in the world?
What are the most important contributing factors to a nation’s happiness?
Is overall happiness increasing or decreasing?